refactor: extract final class (completing the interface), silence Pydantic warnings about shadowed fields
